legal-advisor-platform/backend/requirements.txt
Flask==2.1.0
Flask-CORS==3.0
pandas==1.3.3
google-generativeai==0.3.1
openpyxl==3.0.9

Now, let's update the project's `package.json` to include a script for running both the frontend and backend:

```json file="package.json"
{
  "scripts": {
    "dev": "next dev",
    "build": "next build",
    "start": "next start",
    "lint": "next lint",
    "backend": "cd backend && flask run",
    "dev:all": "concurrently \"npm run dev\" \"npm run backend\""
  },
  "dependencies": {
    // ... existing dependencies ...
    "concurrently": "^6.2.1"
  }
}

